file: Rename files and extract index, target to separate files
